create game loop, add engine control to the system, add SysUpdate (calls only once for system)

+++ src/EntityComponentSystem/include/ECS/Event/EventDispatcher.h
@@ -18,7 +18,7 @@ class EventDispatcher {
       : eventSender(eventSender), eventListener(eventListener) {}
 
   // Invoke all the events from the incoming queue
-  void DispatchEvents(float deltaTime) {
+  void DispatchEvents(double deltaTime) {
       //TODO: is it fast enough? maybe use map?
       for (auto event: eventSender->events) {
           event->update(deltaTime);
